# Query optimization for top reactions.